Adding middleware that checks if user is authenticated, adding the token of the user in the frontend to be send to the backend. Redirecting the user routes in the frontend based of if the user is logged in.

+++ frontend/src/components/WorkoutDetails.jsx
@@ -1,13 +1,22 @@
 import React from "react";
 import { useWorkoutsContext } from "../hooks/useWorkoutsContex";
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = ({ workout }) => {
     console.log(workout)
-    const { dispatch } = useWorkoutsContext()
+    const { dispatch } = useWorkoutsContext();
+    const { user } = useAuthContext();
     const handleClick = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch(`http://localhost:4000/api/workouts/${workout._id}`, {
-            method: "DELETE"
+            method: "DELETE",
+            headers: {
+                "Authorization": `Bearer ${user.token}`
+            }
         })
         const json = await response.json();
         if (response.ok) {
